refactor(bus): streamline planner data and admin flows
