+++ src/features/bus/components/bus-planner-controls.tsx
@@ -1,11 +1,19 @@
 "use client";
 
-import { ArrowLeftRight } from "lucide-react";
+import {
+  ArrowLeftRight,
+  Eye,
+  EyeOff,
+  Map as MapIcon,
+  MapPin,
+} from "lucide-react";
 import {
   DashboardTabToolbarGroup,
   dashboardTabToolbarItemClass,
 } from "@/components/filters/dashboard-tab-toolbar";
+import { Card } from "@/components/ui/card";
 import type { BusTimetableData } from "@/features/bus/lib/bus-types";
+import { Link } from "@/i18n/routing";
 import { cn } from "@/lib/utils";
 
 function PlannerDayTypePills({
@@ -18,16 +26,17 @@ function PlannerDayTypePills({
   t: (key: string) => string;
 }) {
   return (
-    <DashboardTabToolbarGroup className="rounded-xl border-border/70 bg-background/90 p-1">
+    <DashboardTabToolbarGroup className="border-0 bg-muted/35 p-0.5">
       {(["weekday", "weekend"] as const).map((dayType) => (
         <button
           key={dayType}
           type="button"
           aria-pressed={value === dayType}
           onClick={() => onChange(dayType)}
-          className={dashboardTabToolbarItemClass(
-            value === dayType,
-            "min-h-9 rounded-lg px-3 font-medium text-sm",
+          className={cn(
+            dashboardTabToolbarItemClass(value === dayType),
+            "min-h-8 rounded-md px-3 font-medium text-sm shadow-none",
+            value === dayType && "bg-foreground text-background",
           )}
         >
           {t(`dayType.${dayType}`)}
@@ -51,14 +60,9 @@ function StopPicker({
   onSelect: (campusId: number) => void;
 }) {
   return (
-    <section className="space-y-2.5">
-      <p className="text-muted-foreground text-xs uppercase tracking-[0.14em]">
-        {label}
-      </p>
-      <fieldset
-        data-testid={testId}
-        className="grid gap-2 sm:grid-cols-2 lg:grid-cols-1"
-      >
+    <section className="min-w-0 space-y-2">
+      <p className="font-medium text-foreground text-sm">{label}</p>
+      <fieldset data-testid={testId} className="grid">
         <legend className="sr-only">{label}</legend>
         {campuses.map((campus) => {
           const isSelected = selectedId === campus.id;
@@ -69,20 +73,33 @@ function StopPicker({
               aria-pressed={isSelected}
               onClick={() => onSelect(campus.id)}
               className={cn(
-                "flex min-h-11 w-full touch-manipulation items-center justify-between gap-3 rounded-xl border px-3 py-2.5 text-left transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2",
+                "group flex min-h-11 w-full touch-manipulation items-center justify-between gap-3 px-1 py-2.5 text-left transition-colors hover:bg-muted/30 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-inset",
                 isSelected
-                  ? "border-foreground bg-foreground text-background shadow-[0_1px_2px_rgba(15,23,42,0.08)]"
-                  : "border-border/70 bg-background/80 text-foreground hover:border-foreground/25 hover:bg-muted/30",
+                  ? "bg-muted/45 text-foreground"
+                  : "text-muted-foreground",
               )}
             >
-              <span className="truncate text-sm">{campus.namePrimary}</span>
+              <span className="flex min-w-0 items-center gap-2.5">
+                <span
+                  aria-hidden="true"
+                  className={cn(
+                    "flex size-7 shrink-0 items-center justify-center rounded-md transition-colors",
+                    isSelected
+                      ? "bg-background text-foreground"
+                      : "bg-muted/45 text-muted-foreground",
+                  )}
+                >
+                  <MapPin className="size-3.5" />
+                </span>
+                <span className="truncate text-sm">{campus.namePrimary}</span>
+              </span>
               <span
                 aria-hidden="true"
                 className={cn(
-                  "h-2.5 w-2.5 shrink-0 rounded-full border transition-colors",
+                  "size-2.5 shrink-0 rounded-full transition-colors",
                   isSelected
-                    ? "border-background/80 bg-background"
-                    : "border-border/80 bg-transparent",
+                    ? "bg-foreground"
+                    : "bg-muted-foreground/25 group-hover:bg-muted-foreground/45",
                 )}
               />
             </button>
@@ -98,9 +115,7 @@ export function BusPlannerControls({
   endCampusId,
   handleSwap,
   markDirty,
-  selectedDayType,
   setEndCampusId,
-  setSelectedDayType,
   setStartCampusId,
   startCampusId,
   t,
@@ -109,27 +124,15 @@ export function BusPlannerControls({
   endCampusId: number | null;
   handleSwap: () => void;
   markDirty: () => void;
-  selectedDayType: "weekday" | "weekend";
   setEndCampusId: (value: number | null) => void;
-  setSelectedDayType: (value: "weekday" | "weekend") => void;
   setStartCampusId: (value: number | null) => void;
   startCampusId: number | null;
   t: (key: string) => string;
 }) {
   return (
-    <section className="rounded-2xl border border-border/70 bg-card p-4 shadow-sm sm:p-5">
-      <div className="flex flex-col gap-5">
-        <div className="flex min-w-0 flex-1 flex-col gap-3 border-border/50 border-b pb-5">
-          <div className="flex flex-wrap items-center justify-between gap-2">
-            <PlannerDayTypePills
-              value={selectedDayType}
-              onChange={setSelectedDayType}
-              t={t}
-            />
-          </div>
-        </div>
-
-        <div className="grid min-w-0 flex-1 gap-4 lg:grid-cols-[minmax(0,1fr)_auto_minmax(0,1fr)] lg:items-start">
+    <section className="min-w-0 py-4">
+      <Card className="gap-0 rounded-lg border-border/60 bg-card/80 px-3 py-3">
+        <div className="grid min-w-0 grid-cols-[minmax(0,1fr)_2.25rem_minmax(0,1fr)] items-start gap-3 sm:grid-cols-[minmax(0,1fr)_2.5rem_minmax(0,1fr)] sm:gap-4">
           <StopPicker
             testId="bus-start-stop-group"
             label={t("planner.start")}
@@ -146,14 +149,14 @@ export function BusPlannerControls({
             }}
           />
 
-          <div className="flex items-center justify-center lg:items-start lg:pt-9">
+          <div className="flex items-start justify-center pt-8">
             <button
               type="button"
               onClick={handleSwap}
               className={cn(
                 dashboardTabToolbarItemClass(
                   false,
-                  "inline-flex h-10 w-10 items-center justify-center rounded-xl border border-border/70 bg-background text-foreground transition-colors hover:bg-muted/30 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2",
+                  "inline-flex size-9 items-center justify-center rounded-md bg-muted/35 text-foreground transition-colors hover:bg-muted focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 sm:size-10",
                 ),
               )}
               aria-label={t("planner.reverse")}
@@ -179,7 +182,62 @@ export function BusPlannerControls({
             }}
           />
         </div>
-      </div>
+      </Card>
     </section>
+  );
+}
+
+export function BusPlannerSettings({
+  markDirty,
+  selectedDayType,
+  setSelectedDayType,
+  setShowDepartedTrips,
+  showDepartedTrips,
+  t,
+}: {
+  markDirty: () => void;
+  selectedDayType: "weekday" | "weekend";
+  setSelectedDayType: (value: "weekday" | "weekend") => void;
+  setShowDepartedTrips: (updater: (value: boolean) => boolean) => void;
+  showDepartedTrips: boolean;
+  t: (key: string) => string;
+}) {
+  return (
+    <div className="mb-3 flex flex-wrap items-center justify-end gap-2">
+      <PlannerDayTypePills
+        value={selectedDayType}
+        onChange={setSelectedDayType}
+        t={t}
+      />
+      <button
+        type="button"
+        onClick={() => {
+          markDirty();
+          setShowDepartedTrips((value) => !value);
+        }}
+        className={cn(
+          "inline-flex min-h-9 items-center justify-center gap-2 rounded-md px-3 font-medium text-sm transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2",
+          showDepartedTrips
+            ? "bg-foreground text-background hover:bg-foreground/90"
+            : "bg-muted/35 text-foreground hover:bg-muted",
+        )}
+        aria-pressed={showDepartedTrips}
+        aria-label={t("query.showDepartedTrips")}
+      >
+        {showDepartedTrips ? (
+          <Eye aria-hidden="true" className="h-4 w-4" />
+        ) : (
+          <EyeOff aria-hidden="true" className="h-4 w-4" />
+        )}
+        <span>{t("query.showDepartedTrips")}</span>
+      </button>
+      <Link
+        href="/bus-map"
+        className="inline-flex min-h-9 items-center justify-center gap-2 rounded-md bg-muted/35 px-3 font-medium text-foreground text-sm transition-colors hover:bg-muted focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2"
+      >
+        <MapIcon aria-hidden="true" className="h-4 w-4" />
+        <span>{t("transitMap")}</span>
+      </Link>
+    </div>
   );
 }
